create top and header elements
