+++ tests/part_3/test_currency_main_and_catalog_pages.py
@@ -1,7 +1,7 @@
 import pytest
 
 from srv.constants import CURRENCY
-from srv.pages.base_page import BasePage
+from srv.pages.elements.top_element import Top
 from srv.pages.main_page import MainPageElements
 from srv.webdriver_object import WebdriverObject
 
@@ -15,7 +15,7 @@ def test_currency_main_and_catalog_pages(driver, base_url, url, currency):
     else:
         assert driver.title == 'Your Store'
     webdriver_object = WebdriverObject(driver)
-    base_page_top_currency = BasePage.Top.Currency(currency)
+    base_page_top_currency = Top.Currency(currency)
     webdriver_object.wait_visible_element(base_page_top_currency.CURRENCY_BUTTON).click()
     # check currency menu
     webdriver_object.wait_visible_element(base_page_top_currency.CURRENCY_MENU)
